Refactor large screens by extracting dialogs
