+++ lib/screens/customer_detail/full_screen_image_viewer.dart
@@ -1,0 +1,58 @@
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+
+import '../../models/project_media.dart';
+
+class FullScreenImageViewer extends StatelessWidget {
+  final ProjectMedia mediaItem;
+
+  const FullScreenImageViewer({
+    required this.mediaItem,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.black,
+      appBar: AppBar(
+        backgroundColor: Colors.black.withAlpha(128),
+        foregroundColor: Colors.white,
+        title: Text(
+          mediaItem.fileName,
+          style: const TextStyle(color: Colors.white),
+        ),
+        leading: IconButton(
+          icon: const Icon(Icons.close, color: Colors.white),
+          onPressed: () => Navigator.pop(context),
+        ),
+      ),
+      body: Center(
+        child: InteractiveViewer(
+          minScale: 0.5,
+          maxScale: 3.0,
+          child: Image.file(
+            File(mediaItem.filePath),
+            fit: BoxFit.contain,
+            errorBuilder: (context, error, stackTrace) {
+              return const Center(
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Icon(Icons.broken_image, size: 64, color: Colors.white54),
+                    SizedBox(height: 16),
+                    Text(
+                      'Cannot load image',
+                      style: TextStyle(color: Colors.white54),
+                    ),
+                  ],
+                ),
+              );
+            },
+          ),
+        ),
+      ),
+    );
+  }
+}
