+++ lib/screens/customer_detail_screen.dart
@@ -21,7 +21,11 @@ import '../mixins/file_sharing_mixin.dart';
 import '../models/custom_app_data.dart';
 import '../models/inspection_document.dart';
 import 'inspection_viewer_screen.dart';
-import 'inspection_viewer_screen.dart';
+import 'customer_detail/enhanced_communication_dialog.dart';
+import 'customer_detail/media_details_dialog.dart';
+import 'customer_detail/full_screen_image_viewer.dart';
+import 'customer_detail/category_media_screen.dart';
+import 'customer_detail/customer_edit_dialog.dart';
 
 class CustomerDetailScreen extends StatefulWidget {
   final Customer customer;
@@ -5454,7 +5458,7 @@ class _CustomerDetailScreenState extends State<CustomerDetailScreen>
       final ProjectMedia? mediaItem = await showDialog<ProjectMedia>(
         context: context,
         barrierDismissible: false,
-        builder: (context) => _MediaDetailsDialog(
+        builder: (context) => MediaDetailsDialog(
           file: file,
           fileName: fileName,
           fileType: detectedType,
@@ -5492,7 +5496,7 @@ class _CustomerDetailScreenState extends State<CustomerDetailScreen>
         Navigator.push(
           context,
           MaterialPageRoute(
-            builder: (context) => _FullScreenImageViewer(mediaItem: mediaItem),
+            builder: (context) => FullScreenImageViewer(mediaItem: mediaItem),
           ),
         );
       } else if (mediaItem.isPdf) {
@@ -5600,7 +5604,7 @@ class _CustomerDetailScreenState extends State<CustomerDetailScreen>
   void _editMediaDetails(ProjectMedia mediaItem) {
     showDialog(
       context: context,
-      builder: (context) => _MediaDetailsDialog.edit(
+      builder: (context) => MediaDetailsDialog.edit(
         mediaItem: mediaItem,
         onSave: (updatedMedia) async {
           await context.read<AppStateProvider>().updateProjectMedia(updatedMedia);
@@ -5661,7 +5665,7 @@ class _CustomerDetailScreenState extends State<CustomerDetailScreen>
     Navigator.push(
       context,
       MaterialPageRoute(
-        builder: (context) => _CategoryMediaScreen(
+        builder: (context) => CategoryMediaScreen(
           category: category,
           mediaItems: items,
           customerName: widget.customer.name,
@@ -5754,7 +5758,7 @@ class _CustomerDetailScreenState extends State<CustomerDetailScreen>
     showDialog(
       context: context,
       barrierDismissible: false,
-      builder: (context) => _CustomerEditDialog(
+      builder: (context) => CustomerEditDialog(
         customer: widget.customer,
         onCustomerUpdated: () {
           setState(() {});
@@ -5822,7 +5826,7 @@ class _CustomerDetailScreenState extends State<CustomerDetailScreen>
   void _addCommunication() {
     showDialog(
       context: context,
-      builder: (context) => _EnhancedCommunicationDialog(
+      builder: (context) => EnhancedCommunicationDialog(
         customer: widget.customer,
         onCommunicationAdded: () {
           setState(() {}); // Refresh the UI
@@ -6366,1016 +6370,3 @@ class _CustomerDetailScreenState extends State<CustomerDetailScreen>
 
 
 
- // <-- End of CustomerDetailScreen class
-
-// PLACE THE ENHANCED COMMUNICATION DIALOG HERE:
-
-class _EnhancedCommunicationDialog extends StatefulWidget {
-  final Customer customer;
-  final VoidCallback? onCommunicationAdded;
-
-  const _EnhancedCommunicationDialog({
-    required this.customer,
-    this.onCommunicationAdded,
-  });
-
-  @override
-  State<_EnhancedCommunicationDialog> createState() => _EnhancedCommunicationDialogState();
-}
-
-class _EnhancedCommunicationDialogState extends State<_EnhancedCommunicationDialog> {
-  final _formKey = GlobalKey<FormState>();
-  final _contentController = TextEditingController();
-  final _subjectController = TextEditingController();
-
-  String _selectedType = 'call';
-  bool _isUrgent = false;
-
-  final Map<String, Map<String, dynamic>> _communicationTypes = {
-    'call': {'label': '📞 Phone Call', 'hint': 'What was discussed on the call?'},
-    'email': {'label': '📧 Email', 'hint': 'Summary of email content...'},
-    'text': {'label': '💬 Text Message', 'hint': 'Text message content...'},
-  };
-
-  @override
-  void dispose() {
-    _contentController.dispose();
-    _subjectController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-      child: Container(
-        width: MediaQuery.of(context).size.width * 0.9,
-        constraints: const BoxConstraints(maxHeight: 600),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            // Header
-            Container(
-              padding: const EdgeInsets.all(20),
-              decoration: BoxDecoration(
-                color: Theme.of(context).primaryColor.withValues(alpha: 0.1),
-                borderRadius: const BorderRadius.only(
-                  topLeft: Radius.circular(16),
-                  topRight: Radius.circular(16),
-                ),
-              ),
-              child: Row(
-                children: [
-                  Icon(Icons.add_comment, color: Theme.of(context).primaryColor, size: 28),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: Text(
-                      'Add Communication',
-                      style: Theme.of(context).textTheme.titleLarge?.copyWith(fontWeight: FontWeight.bold),
-                    ),
-                  ),
-                  IconButton(icon: const Icon(Icons.close), onPressed: () => Navigator.pop(context)),
-                ],
-              ),
-            ),
-
-            // Content
-            Flexible(
-              child: SingleChildScrollView(
-                padding: const EdgeInsets.all(20),
-                child: Form(
-                  key: _formKey,
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      // Communication Type
-                      Text('Communication Type', style: Theme.of(context).textTheme.titleSmall?.copyWith(fontWeight: FontWeight.bold)),
-                      const SizedBox(height: 8),
-                      DropdownButtonFormField<String>(
-                        value: _selectedType,
-                        decoration: InputDecoration(
-                          border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
-                          contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-                        ),
-                        items: _communicationTypes.entries.map((entry) {
-                          return DropdownMenuItem(
-                            value: entry.key,
-                            child: Text(entry.value['label'] as String),
-                          );
-                        }).toList(),
-                        onChanged: (value) => setState(() => _selectedType = value ?? 'note'),
-                      ),
-
-                      const SizedBox(height: 16),
-
-                      // Subject (for emails/meetings)
-                      if (_selectedType == 'email' || _selectedType == 'meeting') ...[
-                        Text('Subject', style: Theme.of(context).textTheme.titleSmall?.copyWith(fontWeight: FontWeight.bold)),
-                        const SizedBox(height: 8),
-                        TextFormField(
-                          controller: _subjectController,
-                          decoration: InputDecoration(
-                            border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
-                            hintText: _selectedType == 'email' ? 'Email subject line...' : 'Meeting topic...',
-                            prefixIcon: const Icon(Icons.subject),
-                          ),
-                        ),
-                        const SizedBox(height: 16),
-                      ],
-
-                      // Content
-                      Text('Content', style: Theme.of(context).textTheme.titleSmall?.copyWith(fontWeight: FontWeight.bold)),
-                      const SizedBox(height: 8),
-                      TextFormField(
-                        controller: _contentController,
-                        decoration: InputDecoration(
-                          border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
-                          hintText: _communicationTypes[_selectedType]!['hint'] as String,
-                          prefixIcon: const Icon(Icons.notes),
-                        ),
-                        maxLines: 4,
-                        validator: (value) {
-                          if (value == null || value.trim().isEmpty) {
-                            return 'Please enter communication content';
-                          }
-                          return null;
-                        },
-                      ),
-
-                      const SizedBox(height: 16),
-
-                      // Urgent checkbox
-                      CheckboxListTile(
-                        title: const Text('Mark as Urgent'),
-                        subtitle: const Text('High priority communication'),
-                        value: _isUrgent,
-                        onChanged: (value) => setState(() => _isUrgent = value ?? false),
-                        controlAffinity: ListTileControlAffinity.leading,
-                        contentPadding: EdgeInsets.zero,
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-
-            // Footer
-            Container(
-              padding: const EdgeInsets.all(20),
-              decoration: BoxDecoration(
-                color: Colors.grey[50],
-                borderRadius: const BorderRadius.only(
-                  bottomLeft: Radius.circular(16),
-                  bottomRight: Radius.circular(16),
-                ),
-              ),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.end,
-                children: [
-                  OutlinedButton(
-                    onPressed: () => Navigator.pop(context),
-                    child: const Text('Cancel'),
-                  ),
-                  const SizedBox(width: 12),
-                  ElevatedButton(
-                    onPressed: _saveCommunication,
-                    child: const Text('Add Communication'),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _saveCommunication() {
-    if (!_formKey.currentState!.validate()) return;
-
-    try {
-      // Get the emoji prefix for the communication type
-      String prefix = _communicationTypes[_selectedType]!['label'].toString().split(' ')[0];
-
-      // Build the message
-      String message = prefix;
-
-      if (_isUrgent) message += ' [URGENT]';
-
-      if (_subjectController.text.trim().isNotEmpty) {
-        message += ' [${_subjectController.text.trim()}]';
-      }
-
-      message += ' ${_contentController.text.trim()}';
-
-      // Add to customer using existing method
-      widget.customer.addCommunication(message);
-
-      // Update in database
-      context.read<AppStateProvider>().updateCustomer(widget.customer);
-
-      Navigator.pop(context);
-      widget.onCommunicationAdded?.call();
-
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text('Communication logged successfully!'),
-          backgroundColor: Colors.green,
-        ),
-      );
-
-    } catch (e) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: Text('Error saving communication: $e'),
-          backgroundColor: Colors.red,
-        ),
-      );
-    }
-  }
-}
-
-
-
-
-// MEDIA DETAILS DIALOG
-class _MediaDetailsDialog extends StatefulWidget {
-  final File? file;
-  final String? fileName;
-  final String? fileType;
-  final int? fileSize;
-  final String customerId; // Non-nullable
-  final ProjectMedia? mediaItem;
-  final Function(ProjectMedia)? onSave;
-
-  // Default constructor (can remain const if customerId is the only varying part among final fields)
-  const _MediaDetailsDialog({
-    super.key,
-    this.file,
-    this.fileName,
-    this.fileType,
-    this.fileSize,
-    required this.customerId,
-  })  : mediaItem = null,
-        onSave = null;
-
-  // Edit constructor - REMOVE CONST HERE
-  _MediaDetailsDialog.edit({ // <<<< REMOVED 'const'
-    super.key,
-    required this.mediaItem,
-    required this.onSave,
-  })  : file = null,
-        fileName = null,
-        fileType = null,
-        fileSize = null,
-        customerId = mediaItem!.customerId; // This is now valid in a non-const constructor
-
-  @override
-  State<_MediaDetailsDialog> createState() => _MediaDetailsDialogState();
-}
-
-class _MediaDetailsDialogState extends State<_MediaDetailsDialog> {
-  final _formKey = GlobalKey<FormState>();
-  final _descriptionController = TextEditingController();
-  final _tagController = TextEditingController();
-
-  String _selectedCategory = 'general';
-  List<String> _tags = [];
-
-  final List<String> _categories = [
-    'general',
-    'before_photos',
-    'after_photos',
-    'damage_report',
-    'progress_photos',
-    'roofscope_reports',
-    'contracts',
-    'invoices',
-    'permits',
-    'insurance_docs',
-  ];
-
-  bool get _isEditing => widget.mediaItem != null;
-
-  @override
-  void initState() {
-    super.initState();
-    if (_isEditing) {
-      final media = widget.mediaItem!;
-      _descriptionController.text = media.description ?? '';
-      _selectedCategory = media.category;
-      _tags = List.from(media.tags);
-    }
-  }
-
-  @override
-  void dispose() {
-    _descriptionController.dispose();
-    _tagController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      child: SizedBox(
-        width: MediaQuery.of(context).size.width * 0.9,
-        height: MediaQuery.of(context).size.height * 0.8,
-        child: Column(
-          children: [
-            // Header
-            Container(
-              padding: const EdgeInsets.all(16),
-              decoration: BoxDecoration(
-                color: Theme.of(context).primaryColor.withValues(alpha: 0.1),
-                borderRadius: const BorderRadius.only(
-                  topLeft: Radius.circular(12),
-                  topRight: Radius.circular(12),
-                ),
-              ),
-              child: Row(
-                children: [
-                  Icon(
-                    _isEditing ? Icons.edit : Icons.add_photo_alternate,
-                    color: Theme.of(context).primaryColor,
-                  ),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: Text(
-                      _isEditing ? 'Edit Media Details' : 'Add Media Details',
-                      style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                  ),
-                  IconButton(
-                    icon: const Icon(Icons.close),
-                    onPressed: () => Navigator.pop(context),
-                  ),
-                ],
-              ),
-            ),
-
-            // Content
-            Expanded(
-              child: SingleChildScrollView(
-                padding: const EdgeInsets.all(16),
-                child: Form(
-                  key: _formKey,
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      // File preview/info
-                      if (!_isEditing && widget.file != null) ...[
-                        Card(
-                          child: Padding(
-                            padding: const EdgeInsets.all(16),
-                            child: Row(
-                              children: [
-                                Container(
-                                  width: 48,
-                                  height: 48,
-                                  decoration: BoxDecoration(
-                                    color: Colors.grey[200],
-                                    borderRadius: BorderRadius.circular(8),
-                                  ),
-                                  child: widget.fileType == 'image'
-                                      ? ClipRRect(
-                                    borderRadius: BorderRadius.circular(8),
-                                    child: Image.file(
-                                      widget.file!,
-                                      fit: BoxFit.cover,
-                                      errorBuilder: (context, error, stackTrace) {
-                                        return Icon(Icons.broken_image, color: Colors.grey[400]);
-                                      },
-                                    ),
-                                  )
-                                      : Icon(
-                                    widget.fileType == 'pdf'
-                                        ? Icons.picture_as_pdf
-                                        : Icons.insert_drive_file,
-                                    color: Colors.grey[400],
-                                  ),
-                                ),
-                                const SizedBox(width: 12),
-                                Expanded(
-                                  child: Column(
-                                    crossAxisAlignment: CrossAxisAlignment.start,
-                                    children: [
-                                      Text(
-                                        widget.fileName ?? 'Unknown file',
-                                        style: const TextStyle(fontWeight: FontWeight.w500),
-                                        maxLines: 2,
-                                        overflow: TextOverflow.ellipsis,
-                                      ),
-                                      Text(
-                                        widget.fileSize != null
-                                            ? _formatFileSize(widget.fileSize!)
-                                            : 'Unknown size',
-                                        style: TextStyle(
-                                          color: Colors.grey[600],
-                                          fontSize: 12,
-                                        ),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ),
-                        const SizedBox(height: 16),
-                      ],
-
-                      // Category selection
-                      Text(
-                        'Category',
-                        style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 8),
-                      DropdownButtonFormField<String>(
-                        value: _selectedCategory,
-                        decoration: const InputDecoration(
-                          border: OutlineInputBorder(),
-                          prefixIcon: Icon(Icons.category),
-                        ),
-                        items: _categories.map((category) {
-                          return DropdownMenuItem(
-                            value: category,
-                            child: Text(_formatCategoryName(category)),
-                          );
-                        }).toList(),
-                        onChanged: (value) {
-                          setState(() {
-                            _selectedCategory = value ?? 'general';
-                          });
-                        },
-                      ),
-                      const SizedBox(height: 16),
-
-                      // Description
-                      Text(
-                        'Description (Optional)',
-                        style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 8),
-                      TextFormField(
-                        controller: _descriptionController,
-                        decoration: const InputDecoration(
-                          border: OutlineInputBorder(),
-                          prefixIcon: Icon(Icons.description),
-                          hintText: 'Add a description...',
-                        ),
-                        maxLines: 3,
-                      ),
-                      const SizedBox(height: 16),
-
-                      // Tags
-                      Text(
-                        'Tags',
-                        style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 8),
-                      Row(
-                        children: [
-                          Expanded(
-                            child: TextFormField(
-                              controller: _tagController,
-                              decoration: const InputDecoration(
-                                border: OutlineInputBorder(),
-                                prefixIcon: Icon(Icons.tag),
-                                hintText: 'Add a tag and press Enter',
-                              ),
-                              onFieldSubmitted: _addTag,
-                            ),
-                          ),
-                          const SizedBox(width: 8),
-                          IconButton(
-                            onPressed: () => _addTag(_tagController.text),
-                            icon: const Icon(Icons.add),
-                            style: IconButton.styleFrom(
-                              backgroundColor: Theme.of(context).primaryColor,
-                              foregroundColor: Colors.white,
-                            ),
-                          ),
-                        ],
-                      ),
-                      const SizedBox(height: 8),
-                      if (_tags.isNotEmpty)
-                        Wrap(
-                          spacing: 8,
-                          runSpacing: 4,
-                          children: _tags.map((tag) {
-                            return Chip(
-                              label: Text(tag),
-                              deleteIcon: const Icon(Icons.close, size: 16),
-                              onDeleted: () => _removeTag(tag),
-                            );
-                          }).toList(),
-                        ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-
-            // Footer
-            Container(
-              padding: const EdgeInsets.all(16),
-              decoration: BoxDecoration(
-                color: Colors.grey[50],
-                borderRadius: const BorderRadius.only(
-                  bottomLeft: Radius.circular(12),
-                  bottomRight: Radius.circular(12),
-                ),
-              ),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.end,
-                children: [
-                  OutlinedButton(
-                    onPressed: () => Navigator.pop(context),
-                    child: const Text('Cancel'),
-                  ),
-                  const SizedBox(width: 12),
-                  ElevatedButton(
-                    onPressed: _saveMedia,
-                    child: Text(_isEditing ? 'Update' : 'Add Media'),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void _addTag(String tag) {
-    final trimmedTag = tag.trim().toLowerCase();
-    if (trimmedTag.isNotEmpty && !_tags.contains(trimmedTag)) {
-      setState(() {
-        _tags.add(trimmedTag);
-        _tagController.clear();
-      });
-    }
-  }
-
-  void _removeTag(String tag) {
-    setState(() {
-      _tags.remove(tag);
-    });
-  }
-
-  void _saveMedia() {
-    if (_isEditing) {
-      // Update existing media
-      final updatedMedia = widget.mediaItem!;
-      updatedMedia.updateDetails(
-        description: _descriptionController.text.trim().isEmpty
-            ? null
-            : _descriptionController.text.trim(),
-        category: _selectedCategory,
-        tags: _tags,
-      );
-
-      widget.onSave?.call(updatedMedia);
-      Navigator.pop(context);
-    } else {
-      // Create new media
-      final mediaItem = ProjectMedia(
-        customerId: widget.customerId,
-        filePath: widget.file!.path,
-        fileName: widget.fileName!,
-        fileType: widget.fileType!,
-        description: _descriptionController.text.trim().isEmpty
-            ? null
-            : _descriptionController.text.trim(),
-        category: _selectedCategory,
-        tags: _tags,
-        fileSizeBytes: widget.fileSize,
-      );
-
-      Navigator.pop(context, mediaItem);
-    }
-  }
-
-  String _formatCategoryName(String category) {
-    return category
-        .split('_')
-        .map((word) => word[0].toUpperCase() + word.substring(1))
-        .join(' ');
-  }
-
-  String _formatFileSize(int bytes) {
-    if (bytes < 1024) return '$bytes B';
-    if (bytes < 1024 * 1024) return '${(bytes / 1024).toStringAsFixed(1)} KB';
-    if (bytes < 1024 * 1024 * 1024) return '${(bytes / (1024 * 1024)).toStringAsFixed(1)} MB';
-    return '${(bytes / (1024 * 1024 * 1024)).toStringAsFixed(1)} GB';
-  }
-}
-
-// FULL SCREEN IMAGE VIEWER
-class _FullScreenImageViewer extends StatelessWidget {
-  final ProjectMedia mediaItem;
-
-  const _FullScreenImageViewer({
-    required this.mediaItem,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.black,
-      appBar: AppBar(
-        backgroundColor: Colors.black.withValues(alpha: 0.5),
-        foregroundColor: Colors.white,
-        title: Text(
-          mediaItem.fileName,
-          style: const TextStyle(color: Colors.white),
-        ),
-        leading: IconButton(
-          icon: const Icon(Icons.close, color: Colors.white),
-          onPressed: () => Navigator.pop(context),
-        ),
-      ),
-      body: Center(
-        child: InteractiveViewer(
-          minScale: 0.5,
-          maxScale: 3.0,
-          child: Image.file(
-            File(mediaItem.filePath),
-            fit: BoxFit.contain,
-            errorBuilder: (context, error, stackTrace) {
-              return const Center(
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Icon(Icons.broken_image, size: 64, color: Colors.white54),
-                    SizedBox(height: 16),
-                    Text(
-                      'Cannot load image',
-                      style: TextStyle(color: Colors.white54),
-                    ),
-                  ],
-                ),
-              );
-            },
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// CATEGORY MEDIA SCREEN
-class _CategoryMediaScreen extends StatelessWidget {
-  final String category;
-  final List<ProjectMedia> mediaItems;
-  final String customerName;
-
-  const _CategoryMediaScreen({
-    required this.category,
-    required this.mediaItems,
-    required this.customerName,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Text(_formatCategoryName(category)),
-            Text(
-              '$customerName • ${mediaItems.length} items',
-              style: const TextStyle(fontSize: 12, fontWeight: FontWeight.normal),
-            ),
-          ],
-        ),
-        backgroundColor: Theme.of(context).primaryColor,
-        foregroundColor: Colors.white,
-      ),
-      body: GridView.builder(
-        padding: const EdgeInsets.all(16),
-        gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-          crossAxisCount: 2,
-          crossAxisSpacing: 12,
-          mainAxisSpacing: 12,
-          childAspectRatio: 0.8,
-        ),
-        itemCount: mediaItems.length,
-        itemBuilder: (context, index) {
-          final mediaItem = mediaItems[index];
-          return Card(
-            elevation: 2,
-            clipBehavior: Clip.antiAlias,
-            child: InkWell(
-              onTap: () {
-                // Navigate back to detail screen and view media
-                Navigator.pop(context);
-                // TODO: Trigger view media
-              },
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                children: [
-                  Expanded(
-                    child: Container(
-                      color: Colors.grey[200],
-                      child: mediaItem.isImage
-                          ? Image.file(
-                        File(mediaItem.filePath),
-                        fit: BoxFit.cover,
-                        errorBuilder: (context, error, stackTrace) {
-                          return Icon(
-                            Icons.broken_image,
-                            size: 48,
-                            color: Colors.grey[400],
-                          );
-                        },
-                      )
-                          : Icon(
-                        mediaItem.isPdf
-                            ? Icons.picture_as_pdf_outlined
-                            : Icons.insert_drive_file_outlined,
-                        size: 48,
-                        color: Colors.grey[400],
-                      ),
-                    ),
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          mediaItem.fileName,
-                          style: const TextStyle(
-                            fontWeight: FontWeight.w500,
-                            fontSize: 12,
-                          ),
-                          maxLines: 2,
-                          overflow: TextOverflow.ellipsis,
-                        ),
-                        const SizedBox(height: 4),
-                        Text(
-                          DateFormat('MMM dd, yyyy').format(mediaItem.createdAt),
-                          style: TextStyle(
-                            color: Colors.grey[600],
-                            fontSize: 10,
-                          ),
-                        ),
-                        Text(
-                          mediaItem.formattedFileSize,
-                          style: TextStyle(
-                            color: Colors.grey[600],
-                            fontSize: 10,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  String _formatCategoryName(String category) {
-    return category
-        .split('_')
-        .map((word) => word[0].toUpperCase() + word.substring(1))
-        .join(' ');
-  }
-}
-
-// CUSTOMER EDIT DIALOG (kept from original)
-class _CustomerEditDialog extends StatefulWidget {
-  final Customer customer;
-  final VoidCallback? onCustomerUpdated;
-  const _CustomerEditDialog({
-    required this.customer,
-    this.onCustomerUpdated,
-  });
-
-  @override
-  State<_CustomerEditDialog> createState() => _CustomerEditDialogState();
-}
-
-class _CustomerEditDialogState extends State<_CustomerEditDialog> {
-  final _formKey = GlobalKey<FormState>();
-  final _nameController = TextEditingController();
-  final _phoneController = TextEditingController();
-  final _emailController = TextEditingController();
-  final _streetAddressController = TextEditingController();
-  final _cityController = TextEditingController();
-  final _stateController = TextEditingController();
-  final _zipController = TextEditingController();
-  final _notesController = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    _nameController.text = widget.customer.name;
-    _phoneController.text = widget.customer.phone ?? '';
-    _emailController.text = widget.customer.email ?? '';
-    _streetAddressController.text = widget.customer.streetAddress ?? '';
-    _cityController.text = widget.customer.city ?? '';
-    _stateController.text = widget.customer.stateAbbreviation ?? '';
-    _zipController.text = widget.customer.zipCode ?? '';
-    _notesController.text = widget.customer.notes ?? '';
-  }
-
-  @override
-  void dispose() {
-    _nameController.dispose();
-    _phoneController.dispose();
-    _emailController.dispose();
-    _streetAddressController.dispose();
-    _cityController.dispose();
-    _stateController.dispose();
-    _zipController.dispose();
-    _notesController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-      child: Container(
-        width: MediaQuery.of(context).size.width * 0.9,
-        constraints: const BoxConstraints(maxHeight: 650),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Container(
-              padding: const EdgeInsets.all(20),
-              decoration: BoxDecoration(
-                color: Theme.of(context).primaryColor.withValues(alpha: 0.05),
-                borderRadius: const BorderRadius.only(topLeft: Radius.circular(16), topRight: Radius.circular(16)),
-              ),
-              child: Row(
-                children: [
-                  Icon(Icons.edit_note, color: Theme.of(context).primaryColor, size: 28),
-                  const SizedBox(width: 12),
-                  Text('Edit Customer', style: Theme.of(context).textTheme.titleLarge),
-                  const Spacer(),
-                  IconButton(icon: const Icon(Icons.close), onPressed: () => Navigator.pop(context)),
-                ],
-              ),
-            ),
-            Flexible(
-              child: SingleChildScrollView(
-                padding: const EdgeInsets.all(24),
-                child: Form(
-                  key: _formKey,
-                  child: Column(
-                    children: [
-                      _buildTextField(
-                        controller: _nameController,
-                        label: 'Full Name*',
-                        icon: Icons.person,
-                        validator: (value) => (value == null || value.trim().isEmpty) ? 'Name is required' : null,
-                      ),
-                      const SizedBox(height: 16),
-                      _buildTextField(
-                        controller: _phoneController,
-                        label: 'Phone Number',
-                        icon: Icons.phone,
-                        keyboardType: TextInputType.phone,
-                      ),
-                      const SizedBox(height: 16),
-                      _buildTextField(
-                        controller: _emailController,
-                        label: 'Email Address',
-                        icon: Icons.email,
-                        keyboardType: TextInputType.emailAddress,
-                        validator: (value) {
-                          if (value != null && value.isNotEmpty) {
-                            final emailRegex = RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$');
-                            if (!emailRegex.hasMatch(value)) return 'Enter a valid email address';
-                          }
-                          return null;
-                        },
-                      ),
-                      const SizedBox(height: 16),
-                      const SizedBox(height: 20),
-                      Text("Address Details:", style: Theme.of(context).textTheme.titleSmall?.copyWith(color: Colors.grey[700])),
-                      const Divider(height: 10),
-                      const SizedBox(height: 10),
-                      _buildTextField(controller: _streetAddressController, label: 'Street Address', icon: Icons.home_outlined, hint: "e.g., 123 Main St"),
-                      const SizedBox(height: 16),
-                      Row(
-                        children: [
-                          Expanded(child: _buildTextField(controller: _cityController, label: 'City', icon: Icons.location_city)),
-                          const SizedBox(width: 12),
-                          Expanded(child: _buildTextField(controller: _stateController, label: 'State', icon: Icons.map_outlined, hint: "e.g., WA")),
-                        ],
-                      ),
-                      const SizedBox(height: 16),
-                      _buildTextField(controller: _zipController, label: 'Zip Code', icon: Icons.markunread_mailbox_outlined, keyboardType: TextInputType.number, hint: "e.g., 98001"),
-                      const SizedBox(height: 20),
-                      _buildTextField(
-                        controller: _notesController,
-                        label: 'Notes',
-                        icon: Icons.note_alt_outlined,
-                        maxLines: 3,
-                        hint: 'Additional information...',
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-            Container(
-              padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-              decoration: BoxDecoration(
-                color: Colors.grey[100],
-                borderRadius: const BorderRadius.only(bottomLeft: Radius.circular(16), bottomRight: Radius.circular(16)),
-              ),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.end,
-                children: [
-                  OutlinedButton(onPressed: () => Navigator.pop(context), child: const Text('Cancel')),
-                  const SizedBox(width: 12),
-                  ElevatedButton(onPressed: _saveCustomer, child: const Text('Update Customer')),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildTextField({
-    required TextEditingController controller,
-    required String label,
-    required IconData icon,
-    String? hint,
-    int maxLines = 1,
-    TextInputType? keyboardType,
-    String? Function(String?)? validator,
-  }) {
-    return TextFormField(
-      controller: controller,
-      decoration: InputDecoration(
-        labelText: label,
-        hintText: hint,
-        prefixIcon: Icon(icon, color: Theme.of(context).primaryColor.withValues(alpha: 0.7)),
-        border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
-        focusedBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(12),
-          borderSide: BorderSide(color: Theme.of(context).primaryColor, width: 1.5),
-        ),
-        filled: true,
-        fillColor: Colors.white,
-        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 14),
-      ),
-      maxLines: maxLines,
-      keyboardType: keyboardType,
-      validator: validator,
-    );
-  }
-
-
-
-  void _saveCustomer() {
-    if (!_formKey.currentState!.validate()) return;
-
-    final appState = context.read<AppStateProvider>();
-
-    widget.customer.updateInfo(
-      name: _nameController.text.trim(),
-      phone: _phoneController.text.trim().isEmpty ? null : _phoneController.text.trim(),
-      email: _emailController.text.trim().isEmpty ? null : _emailController.text.trim(),
-      streetAddress: _streetAddressController.text.trim().isEmpty ? null : _streetAddressController.text.trim(),
-      city: _cityController.text.trim().isEmpty ? null : _cityController.text.trim(),
-      stateAbbreviation: _stateController.text.trim().isEmpty ? null : _stateController.text.trim(),
-      zipCode: _zipController.text.trim().isEmpty ? null : _zipController.text.trim(),
-      notes: _notesController.text.trim().isEmpty ? null : _notesController.text.trim(),
-    );
-
-    appState.updateCustomer(widget.customer);
-
-    Navigator.pop(context);
-    widget.onCustomerUpdated?.call();
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(content: Text('Customer updated successfully!'), backgroundColor: Colors.green),
-    );
-  }
-
-
-}
