+++ lib/screens/settings_screen.dart
@@ -13,6 +13,10 @@ import '../providers/app_state_provider.dart';
 import '../services/database_service.dart';
 import '../models/app_settings.dart';
 
+import 'settings/category_manager_dialog.dart';
+import 'settings/units_manager_dialog.dart';
+import 'settings/quote_levels_manager_dialog.dart';
+import 'settings/discount_settings_dialog.dart';
 class SettingsScreen extends StatefulWidget {
   const SettingsScreen({super.key});
 
@@ -1501,7 +1505,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
 
     showDialog(
       context: context,
-      builder: (context) => _CategoryManagerDialog(
+      builder: (context) => CategoryManagerDialog(
         categories: List.from(settings.productCategories),
         onSave: (updatedCategories) {
           settings.updateProductCategories(updatedCategories);
@@ -1525,7 +1529,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
 
     showDialog(
       context: context,
-      builder: (context) => _UnitsManagerDialog(
+      builder: (context) => UnitsManagerDialog(
         units: List.from(settings.productUnits),
         defaultUnit: settings.defaultUnit,
         onSave: (updatedUnits, newDefaultUnit) {
@@ -1551,7 +1555,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
 
     showDialog(
       context: context,
-      builder: (context) => _QuoteLevelsManagerDialog(
+      builder: (context) => QuoteLevelsManagerDialog(
         levelNames: List.from(settings.defaultQuoteLevelNames),
         onSave: (updatedLevels) {
           settings.updateDefaultQuoteLevelNames(updatedLevels);
@@ -1575,7 +1579,7 @@ class _SettingsScreenState extends State<SettingsScreen> {
 
     showDialog(
       context: context,
-      builder: (context) => _DiscountSettingsDialog(
+      builder: (context) => DiscountSettingsDialog(
         discountTypes: List.from(settings.discountTypes),
         defaultDiscountLimit: settings.defaultDiscountLimit,
         onSave: (types, limit) {
@@ -1591,889 +1595,6 @@ class _SettingsScreenState extends State<SettingsScreen> {
           );
         },
       ),
-    );
-  }
-}
-
-// Dialog classes (keeping the existing implementations but with modern styling)
-// Replace the _CategoryManagerDialog class in your settings_screen.dart with this enhanced version
-
-class _CategoryManagerDialog extends StatefulWidget {
-  final List<String> categories;
-  final Function(List<String>) onSave;
-
-  const _CategoryManagerDialog({
-    required this.categories,
-    required this.onSave,
-  });
-
-  @override
-  State<_CategoryManagerDialog> createState() => _CategoryManagerDialogState();
-}
-
-class _CategoryManagerDialogState extends State<_CategoryManagerDialog> {
-  late List<String> _categories;
-  final TextEditingController _addController = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    _categories = List.from(widget.categories);
-  }
-
-  @override
-  void dispose() {
-    _addController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-      title: Row(
-        children: [
-          Container(
-            padding: const EdgeInsets.all(8),
-            decoration: BoxDecoration(
-              color: Colors.blue.shade100,
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Icon(Icons.category, color: Colors.blue.shade600),
-          ),
-          const SizedBox(width: 12),
-          const Text('Manage Product Categories'),
-        ],
-      ),
-      content: SizedBox(
-        width: double.maxFinite,
-        height: 450, // Increased height to accommodate edit functionality
-        child: Column(
-          children: [
-            // Add new category section
-            Container(
-              padding: const EdgeInsets.all(12),
-              decoration: BoxDecoration(
-                color: Colors.blue.shade50,
-                borderRadius: BorderRadius.circular(8),
-                border: Border.all(color: Colors.blue.shade200),
-              ),
-              child: Row(
-                children: [
-                  Expanded(
-                    child: TextField(
-                      controller: _addController,
-                      decoration: InputDecoration(
-                        labelText: 'New Category',
-                        border: OutlineInputBorder(borderRadius: BorderRadius.circular(8)),
-                        isDense: true,
-                        filled: true,
-                        fillColor: Colors.white,
-                        prefixIcon: Icon(Icons.add, color: Colors.blue.shade600),
-                      ),
-                      onSubmitted: (_) => _addCategory(),
-                    ),
-                  ),
-                  const SizedBox(width: 12),
-                  ElevatedButton(
-                    onPressed: _addCategory,
-                    style: ElevatedButton.styleFrom(
-                      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-                      backgroundColor: Colors.blue.shade600,
-                      foregroundColor: Colors.white,
-                    ),
-                    child: const Icon(Icons.add),
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 16),
-
-            // Categories list header
-            Container(
-              padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-              child: Row(
-                children: [
-                  Text(
-                    'Current Categories (${_categories.length})',
-                    style: TextStyle(
-                      fontWeight: FontWeight.w600,
-                      color: Colors.grey[700],
-                    ),
-                  ),
-                  const Spacer(),
-                  if (_categories.length > 1)
-                    Text(
-                      'Tap to edit',
-                      style: TextStyle(
-                        fontSize: 12,
-                        color: Colors.grey[500],
-                      ),
-                    ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 8),
-
-            // Categories list
-            Expanded(
-              child: ListView.builder(
-                itemCount: _categories.length,
-                itemBuilder: (context, index) {
-                  final category = _categories[index];
-                  return Container(
-                    margin: const EdgeInsets.only(bottom: 8),
-                    child: Card(
-                      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-                      child: ListTile(
-                        leading: CircleAvatar(
-                          backgroundColor: Colors.blue.shade100,
-                          radius: 20,
-                          child: Icon(
-                            Icons.category,
-                            color: Colors.blue.shade600,
-                            size: 20,
-                          ),
-                        ),
-                        title: Text(
-                          category,
-                          style: const TextStyle(
-                            fontWeight: FontWeight.w500,
-                          ),
-                        ),
-                        trailing: Row(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            // Edit button
-                            IconButton(
-                              icon: Icon(Icons.edit_outlined, color: Colors.blue.shade600),
-                              onPressed: () => _editCategory(index, category),
-                              tooltip: 'Edit category',
-                            ),
-                            // Delete button (only show if more than 1 category)
-                            if (_categories.length > 1)
-                              IconButton(
-                                icon: Icon(Icons.delete_outline, color: Colors.red.shade600),
-                                onPressed: () => _removeCategory(index),
-                                tooltip: 'Delete category',
-                              ),
-                          ],
-                        ),
-                        onTap: () => _editCategory(index, category),
-                      ),
-                    ),
-                  );
-                },
-              ),
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context),
-          child: const Text('Cancel'),
-        ),
-        ElevatedButton(
-          onPressed: () {
-            widget.onSave(_categories);
-            Navigator.pop(context);
-          },
-          child: const Text('Save Changes'),
-        ),
-      ],
-    );
-  }
-
-  void _addCategory() {
-    final newCategory = _addController.text.trim();
-    if (newCategory.isNotEmpty && !_categories.contains(newCategory)) {
-      setState(() {
-        _categories.add(newCategory);
-        _addController.clear();
-      });
-    } else if (_categories.contains(newCategory)) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: const Text('Category already exists'),
-          backgroundColor: Colors.orange,
-          behavior: SnackBarBehavior.floating,
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-        ),
-      );
-    }
-  }
-
-  void _removeCategory(int index) {
-    if (_categories.length > 1) {
-      showDialog(
-        context: context,
-        builder: (context) => AlertDialog(
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
-          title: Row(
-            children: [
-              Icon(Icons.warning_amber_rounded, color: Colors.orange.shade600),
-              const SizedBox(width: 12),
-              const Text('Delete Category'),
-            ],
-          ),
-          content: Text(
-            'Are you sure you want to delete "${_categories[index]}"?\n\nProducts in this category will need to be reassigned.',
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.pop(context),
-              child: const Text('Cancel'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.pop(context);
-                setState(() {
-                  _categories.removeAt(index);
-                });
-              },
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.red.shade600,
-                foregroundColor: Colors.white,
-              ),
-              child: const Text('Delete'),
-            ),
-          ],
-        ),
-      );
-    }
-  }
-
-  void _editCategory(int index, String currentName) {
-    final TextEditingController editController = TextEditingController(text: currentName);
-
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-        title: Row(
-          children: [
-            Container(
-              padding: const EdgeInsets.all(8),
-              decoration: BoxDecoration(
-                color: Colors.blue.shade100,
-                borderRadius: BorderRadius.circular(8),
-              ),
-              child: Icon(Icons.edit, color: Colors.blue.shade600),
-            ),
-            const SizedBox(width: 12),
-            const Text('Edit Category'),
-          ],
-        ),
-        content: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            TextField(
-              controller: editController,
-              decoration: InputDecoration(
-                labelText: 'Category Name',
-                prefixIcon: Icon(Icons.category, color: Colors.blue.shade600),
-                border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
-                focusedBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(12),
-                  borderSide: BorderSide(color: Colors.blue.shade600, width: 2),
-                ),
-                filled: true,
-                fillColor: Colors.white,
-              ),
-              autofocus: true,
-              onSubmitted: (value) {
-                if (value.trim().isNotEmpty) {
-                  _updateCategory(index, value.trim(), editController);
-                }
-              },
-            ),
-            const SizedBox(height: 16),
-            Container(
-              padding: const EdgeInsets.all(12),
-              decoration: BoxDecoration(
-                color: Colors.blue.shade50,
-                borderRadius: BorderRadius.circular(8),
-                border: Border.all(color: Colors.blue.shade200),
-              ),
-              child: Row(
-                children: [
-                  Icon(Icons.info_outline, color: Colors.blue.shade600, size: 20),
-                  const SizedBox(width: 8),
-                  const Expanded(
-                    child: Text(
-                      'Existing products with this category will be updated automatically.',
-                      style: TextStyle(fontSize: 12),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-        actions: [
-          TextButton(
-            onPressed: () {
-              editController.dispose();
-              Navigator.pop(context);
-            },
-            child: const Text('Cancel'),
-          ),
-          ElevatedButton(
-            onPressed: () {
-              final newName = editController.text.trim();
-              if (newName.isNotEmpty) {
-                _updateCategory(index, newName, editController);
-              }
-            },
-            style: ElevatedButton.styleFrom(
-              backgroundColor: Colors.blue.shade600,
-              foregroundColor: Colors.white,
-            ),
-            child: const Text('Update'),
-          ),
-        ],
-      ),
-    );
-  }
-
-  void _updateCategory(int index, String newName, TextEditingController controller) {
-    if (newName == _categories[index]) {
-      // No change
-      controller.dispose();
-      Navigator.pop(context);
-      return;
-    }
-
-    if (_categories.contains(newName)) {
-      // Category name already exists
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(
-          content: const Text('Category name already exists'),
-          backgroundColor: Colors.orange,
-          behavior: SnackBarBehavior.floating,
-          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-        ),
-      );
-      return;
-    }
-
-    setState(() {
-      _categories[index] = newName;
-    });
-
-    controller.dispose();
-    Navigator.pop(context);
-
-    ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(
-        content: Text('Category updated to "$newName"'),
-        backgroundColor: Colors.green,
-        behavior: SnackBarBehavior.floating,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-      ),
-    );
-  }
-}
-
-class _UnitsManagerDialog extends StatefulWidget {
-  final List<String> units;
-  final String defaultUnit;
-  final Function(List<String>, String) onSave;
-
-  const _UnitsManagerDialog({
-    required this.units,
-    required this.defaultUnit,
-    required this.onSave,
-  });
-
-  @override
-  State<_UnitsManagerDialog> createState() => _UnitsManagerDialogState();
-}
-
-class _UnitsManagerDialogState extends State<_UnitsManagerDialog> {
-  late List<String> _units;
-  late String _defaultUnit;
-  final TextEditingController _addController = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    _units = List.from(widget.units);
-    _defaultUnit = widget.defaultUnit;
-  }
-
-  @override
-  void dispose() {
-    _addController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-      title: Row(
-        children: [
-          Container(
-            padding: const EdgeInsets.all(8),
-            decoration: BoxDecoration(
-              color: Colors.green.shade100,
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Icon(Icons.straighten, color: Colors.green.shade600),
-          ),
-          const SizedBox(width: 12),
-          const Text('Manage Product Units'),
-        ],
-      ),
-      content: SizedBox(
-        width: double.maxFinite,
-        height: 400,
-        child: Column(
-          children: [
-            Container(
-              padding: const EdgeInsets.all(12),
-              decoration: BoxDecoration(
-                color: Colors.green.shade50,
-                borderRadius: BorderRadius.circular(8),
-                border: Border.all(color: Colors.green.shade200),
-              ),
-              child: Column(
-                children: [
-                  Row(
-                    children: [
-                      Expanded(
-                        child: TextField(
-                          controller: _addController,
-                          decoration: InputDecoration(
-                            labelText: 'New Unit',
-                            border: OutlineInputBorder(borderRadius: BorderRadius.circular(8)),
-                            isDense: true,
-                            filled: true,
-                            fillColor: Colors.white,
-                          ),
-                          onSubmitted: (_) => _addUnit(),
-                        ),
-                      ),
-                      const SizedBox(width: 12),
-                      ElevatedButton(
-                        onPressed: _addUnit,
-                        style: ElevatedButton.styleFrom(
-                          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-                        ),
-                        child: const Icon(Icons.add),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 12),
-                  DropdownButtonFormField<String>(
-                    value: _defaultUnit,
-                    decoration: InputDecoration(
-                      labelText: 'Default Unit',
-                      border: OutlineInputBorder(borderRadius: BorderRadius.circular(8)),
-                      isDense: true,
-                      filled: true,
-                      fillColor: Colors.white,
-                    ),
-                    items: _units.map((unit) => DropdownMenuItem(
-                      value: unit,
-                      child: Text(unit),
-                    )).toList(),
-                    onChanged: (value) {
-                      if (value != null) {
-                        setState(() {
-                          _defaultUnit = value;
-                        });
-                      }
-                    },
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 16),
-            Expanded(
-              child: ListView.builder(
-                itemCount: _units.length,
-                itemBuilder: (context, index) {
-                  final unit = _units[index];
-                  final isDefault = unit == _defaultUnit;
-                  return Container(
-                    margin: const EdgeInsets.only(bottom: 8),
-                    child: Card(
-                      color: isDefault ? Colors.green.shade50 : null,
-                      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-                      child: ListTile(
-                        leading: Icon(
-                          isDefault ? Icons.star : Icons.straighten,
-                          color: isDefault ? Colors.green.shade700 : Colors.grey[600],
-                        ),
-                        title: Text(
-                          unit,
-                          style: TextStyle(
-                            fontWeight: isDefault ? FontWeight.w600 : FontWeight.normal,
-                          ),
-                        ),
-                        subtitle: isDefault ? const Text('Default unit') : null,
-                        trailing: _units.length > 1 && !isDefault
-                            ? IconButton(
-                          icon: Icon(Icons.delete_outline, color: Colors.red.shade600),
-                          onPressed: () => _removeUnit(index),
-                        )
-                            : null,
-                      ),
-                    ),
-                  );
-                },
-              ),
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context),
-          child: const Text('Cancel'),
-        ),
-        ElevatedButton(
-          onPressed: () {
-            widget.onSave(_units, _defaultUnit);
-            Navigator.pop(context);
-          },
-          child: const Text('Save'),
-        ),
-      ],
-    );
-  }
-
-  void _addUnit() {
-    final newUnit = _addController.text.trim();
-    if (newUnit.isNotEmpty && !_units.contains(newUnit)) {
-      setState(() {
-        _units.add(newUnit);
-        _addController.clear();
-      });
-    }
-  }
-
-  void _removeUnit(int index) {
-    if (_units.length > 1 && _units[index] != _defaultUnit) {
-      setState(() {
-        _units.removeAt(index);
-      });
-    }
-  }
-}
-
-class _QuoteLevelsManagerDialog extends StatefulWidget {
-  final List<String> levelNames;
-  final Function(List<String>) onSave;
-
-  const _QuoteLevelsManagerDialog({
-    required this.levelNames,
-    required this.onSave,
-  });
-
-  @override
-  State<_QuoteLevelsManagerDialog> createState() => _QuoteLevelsManagerDialogState();
-}
-
-class _QuoteLevelsManagerDialogState extends State<_QuoteLevelsManagerDialog> {
-  late List<TextEditingController> _controllers;
-
-  @override
-  void initState() {
-    super.initState();
-    _controllers = widget.levelNames.map((name) => TextEditingController(text: name)).toList();
-    while (_controllers.length < 3) {
-      _controllers.add(TextEditingController());
-    }
-  }
-
-  @override
-  void dispose() {
-    for (final controller in _controllers) {
-      controller.dispose();
-    }
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-      title: Row(
-        children: [
-          Container(
-            padding: const EdgeInsets.all(8),
-            decoration: BoxDecoration(
-              color: Colors.purple.shade100,
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Icon(Icons.layers, color: Colors.purple.shade600),
-          ),
-          const SizedBox(width: 12),
-          const Text('Configure Quote Levels'),
-        ],
-      ),
-      content: SizedBox(
-        width: double.maxFinite,
-        height: 350,
-        child: Column(
-          children: [
-            Container(
-              padding: const EdgeInsets.all(12),
-              decoration: BoxDecoration(
-                color: Colors.purple.shade50,
-                borderRadius: BorderRadius.circular(8),
-                border: Border.all(color: Colors.purple.shade200),
-              ),
-              child: Row(
-                children: [
-                  Icon(Icons.info_outline, color: Colors.purple.shade600, size: 20),
-                  const SizedBox(width: 8),
-                  const Expanded(
-                    child: Text(
-                      'Set names for default quote levels used in product pricing.',
-                      style: TextStyle(fontSize: 13),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            const SizedBox(height: 16),
-            Expanded(
-              child: ListView.builder(
-                itemCount: _controllers.length,
-                itemBuilder: (context, index) {
-                  return Container(
-                    margin: const EdgeInsets.only(bottom: 12),
-                    child: Row(
-                      children: [
-                        Container(
-                          width: 32,
-                          height: 32,
-                          decoration: BoxDecoration(
-                            color: Colors.purple.shade100,
-                            borderRadius: BorderRadius.circular(8),
-                          ),
-                          child: Center(
-                            child: Text(
-                              '${index + 1}',
-                              style: TextStyle(
-                                fontWeight: FontWeight.bold,
-                                color: Colors.purple.shade700,
-                              ),
-                            ),
-                          ),
-                        ),
-                        const SizedBox(width: 12),
-                        Expanded(
-                          child: TextField(
-                            controller: _controllers[index],
-                            decoration: InputDecoration(
-                              labelText: 'Level ${index + 1} Name',
-                              border: OutlineInputBorder(borderRadius: BorderRadius.circular(8)),
-                              filled: true,
-                              fillColor: Colors.white,
-                              isDense: true,
-                            ),
-                          ),
-                        ),
-                        if (_controllers.length > 3) ...[
-                          const SizedBox(width: 8),
-                          IconButton(
-                            icon: Icon(Icons.remove_circle_outline, color: Colors.red.shade600),
-                            onPressed: () => _removeLevel(index),
-                          ),
-                        ],
-                      ],
-                    ),
-                  );
-                },
-              ),
-            ),
-            ElevatedButton.icon(
-              onPressed: _addLevel,
-              icon: const Icon(Icons.add),
-              label: const Text('Add Level'),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.purple.shade600,
-                foregroundColor: Colors.white,
-              ),
-            ),
-          ],
-        ),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context),
-          child: const Text('Cancel'),
-        ),
-        ElevatedButton(
-          onPressed: () {
-            final levelNames = _controllers
-                .map((c) => c.text.trim())
-                .where((name) => name.isNotEmpty)
-                .toList();
-            widget.onSave(levelNames);
-            Navigator.pop(context);
-          },
-          child: const Text('Save'),
-        ),
-      ],
-    );
-  }
-
-  void _addLevel() {
-    setState(() {
-      _controllers.add(TextEditingController());
-    });
-  }
-
-  void _removeLevel(int index) {
-    if (_controllers.length > 3) {
-      setState(() {
-        _controllers[index].dispose();
-        _controllers.removeAt(index);
-      });
-    }
-  }
-}
-
-class _DiscountSettingsDialog extends StatefulWidget {
-  final List<String> discountTypes;
-  final double defaultDiscountLimit;
-  final Function(List<String>, double) onSave;
-
-  const _DiscountSettingsDialog({
-    required this.discountTypes,
-    required this.defaultDiscountLimit,
-    required this.onSave,
-  });
-
-  @override
-  State<_DiscountSettingsDialog> createState() => _DiscountSettingsDialogState();
-}
-
-class _DiscountSettingsDialogState extends State<_DiscountSettingsDialog> {
-  late List<String> _discountTypes;
-  late double _discountLimit;
-  final TextEditingController _limitController = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    _discountTypes = List.from(widget.discountTypes);
-    _discountLimit = widget.defaultDiscountLimit;
-    _limitController.text = _discountLimit.toStringAsFixed(1);
-  }
-
-  @override
-  void dispose() {
-    _limitController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
-      title: Row(
-        children: [
-          Container(
-            padding: const EdgeInsets.all(8),
-            decoration: BoxDecoration(
-              color: Colors.orange.shade100,
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Icon(Icons.discount, color: Colors.orange.shade600),
-          ),
-          const SizedBox(width: 12),
-          const Text('Discount Settings'),
-        ],
-      ),
-      content: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          TextField(
-            controller: _limitController,
-            decoration: InputDecoration(
-              labelText: 'Maximum Discount Percentage',
-              prefixIcon: Icon(Icons.percent, color: Colors.orange.shade600),
-              suffixText: '%',
-              border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
-              filled: true,
-              fillColor: Colors.white,
-            ),
-            keyboardType: const TextInputType.numberWithOptions(decimal: true),
-            onChanged: (value) {
-              _discountLimit = double.tryParse(value) ?? _discountLimit;
-            },
-          ),
-          const SizedBox(height: 20),
-          Container(
-            width: double.infinity,
-            padding: const EdgeInsets.all(12),
-            decoration: BoxDecoration(
-              color: Colors.orange.shade50,
-              borderRadius: BorderRadius.circular(8),
-              border: Border.all(color: Colors.orange.shade200),
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  'Available Discount Types:',
-                  style: TextStyle(
-                    fontWeight: FontWeight.w600,
-                    color: Colors.orange.shade800,
-                  ),
-                ),
-                const SizedBox(height: 12),
-                ...['percentage', 'fixed_amount', 'voucher'].map((type) {
-                  return Container(
-                    margin: const EdgeInsets.only(bottom: 4),
-                    child: CheckboxListTile(
-                      title: Text(
-                        type.replaceAll('_', ' ').toUpperCase(),
-                        style: const TextStyle(fontSize: 14),
-                      ),
-                      value: _discountTypes.contains(type),
-                      onChanged: (value) {
-                        setState(() {
-                          if (value == true) {
-                            if (!_discountTypes.contains(type)) {
-                              _discountTypes.add(type);
-                            }
-                          } else {
-                            _discountTypes.remove(type);
-                          }
-                        });
-                      },
-                      activeColor: Colors.orange.shade600,
-                      dense: true,
-                    ),
-                  );
-                }),
-              ],
-            ),
-          ),
-        ],
-      ),
-      actions: [
-        TextButton(
-          onPressed: () => Navigator.pop(context),
-          child: const Text('Cancel'),
-        ),
-        ElevatedButton(
-          onPressed: () {
-            widget.onSave(_discountTypes, _discountLimit);
-            Navigator.pop(context);
-          },
-          child: const Text('Save'),
-        ),
-      ],
     );
   }
 }
